+++ src/Electron/Remote.ts
@@ -7,7 +7,7 @@ import {GA} from '../Util/GA';
 // import {StreamEmitter} from '../Main/Stream/StreamEmitter';
 // import {StreamLauncher} from '../Main/Stream/StreamLauncher';
 // import {SystemStreamLauncher} from '../Main/Stream/SystemStreamLauncher';
-import {GitHubClient} from '../Main/GitHub/GitHubClient';
+// import {GitHubClient} from '../Main/GitHub/GitHubClient';
 
 // export const RemoteDB: typeof DB = remote.require('./Main/DB/DB').DB;
 // export const RemoteLogger = remote.require('color-logger').default;
@@ -15,7 +15,7 @@ import {GitHubClient} from '../Main/GitHub/GitHubClient';
 // export const RemoteSystemStreamLauncher: typeof SystemStreamLauncher = remote.require('./Main/Stream/SystemStreamLauncher').SystemStreamLauncher;
 // export const RemoteStreamEmitter: typeof StreamEmitter = remote.require('./Main/Stream/StreamEmitter').StreamEmitter;
 export const RemoteConfig: typeof Config = remote.require('./Main/Config').Config;
-export const RemoteGitHubClient: typeof GitHubClient = remote.require('./Main/GitHub/GitHubClient').GitHubClient;
+// export const RemoteGitHubClient: typeof GitHubClient = remote.require('./Main/GitHub/GitHubClient').GitHubClient;
 export const RemoteGA: typeof GA = remote.require('./Util/GA').GA;
 // export const RemoteDateConverter: typeof DateConverter = remote.require('./Util/DateConverter').DateConverter;
 export const RemoteBrowserViewProxy: typeof BrowserViewProxy = remote.require('./Main/BrowserViewProxy').BrowserViewProxy;
